Revert "PMM-7781 Run IA API tests in parallel"
This reverts commit 258e74025205063d28fe7a844564647ee1bc3bc4.

+++ services/management/ia/common.go
@@ -168,7 +168,7 @@ func validateParameters(definitions models.ParamsDefinitions, values models.Para
 		switch d.Type {
 		case models.Float:
 			v := d.FloatParam
-			fv := value.FloatValue
+			fv := float64(value.FloatValue)
 			if v.Min != nil && pointer.GetFloat64(v.Min) > fv {
 				return status.Errorf(codes.InvalidArgument, "Parameter %s value is less than required minimum.", d.Name)
 			}
